move span-context to a separate file
